Updated: Models to use hierarchical structure
It will be easier to look up the game state if the Game moves are stored in
a tree structure.
The first move will be represented by the tree root and subsequent moves can be
tracked from there till the leaves.

Inserts/updating may be slower but in a 2 player game (where one user is human)
this should be negligible.
Reads are faster which should help performance of minimax.